+++ modules/portio.c
@@ -11,7 +11,7 @@
 #include <kernel/fs.h>
 #include <kernel/module.h>
 
-static uint32_t read_port(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
+static uint32_t read_port(fs_node_t *node, uint64_t offset, uint32_t size, uint8_t *buffer) {
 	switch (size) {
 		case 1:
 			buffer[0] = inportb(offset);
@@ -32,7 +32,7 @@ static uint32_t read_port(fs_node_t *node, uint32_t offset, uint32_t size, uint8
 	return size;
 }
 
-static uint32_t write_port(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer) {
+static uint32_t write_port(fs_node_t *node, uint64_t offset, uint32_t size, uint8_t *buffer) {
 	switch (size) {
 		case 1:
 			outportb(offset, buffer[0]);
